+++ src/js/layout-basic.js
@@ -1,13 +1,23 @@
-const posterkit = require('posterkit.js');
+// -*- coding: utf-8 -*-
+// (c) 2018 The PosterKit developers <developers@posterkit.org>
 require('version.js');
-
+const posterkit = require('posterkit.js');
 
 $(document).ready(function() {
+
+    // Hello world
+    posterkit.welcome();
 
     // Propagate version
     $('#version').html(__version__);
 
-    // Resize all texts with class="fit" to fit their parent containers
-    posterkit.fit_text('.fit');
+    // Read URL parameters and compute runtime settings
+    var options = posterkit.get_runtime_settings();
+
+    // Setup display
+    posterkit.setup_display(options);
+
+    // Run automatic layout engine
+    posterkit.run_autolayout();
 
 });